requirements.txt
booleano==1.1a1
Django==1.11.3
requests==2.18.1
pandas==0.21.1
beautifulsoup4==4.6.0
python_dateutil==2.6.1
mysqlclient
